requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP client for fetching presentations
httpx>=0.25.0

# Conversion Dependencies
playwright>=1.40.0
python-pptx>=0.6.23
Pillow>=10.0.0
